dataHandler 獨立拆分出來
