+++ frontend/src/utils/fetchEnvironment.js
