+++ frontend/src/utils/fetchSafety.js
